+++ akshare/fund/fund_rating.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding:utf-8 -*-
 """
-Date: 2021/3/29 17:48
+Date: 2022/8/8 14:48
 Desc: 基金评级
 http://fund.eastmoney.com/data/fundrating.html
 """
@@ -20,7 +20,9 @@ def fund_rating_all() -> pd.DataFrame:
     url = "http://fund.eastmoney.com/data/fundrating.html"
     r = requests.get(url)
     soup = BeautifulSoup(r.text, "lxml")
-    data_text = soup.find("div", attrs={"id": "fundinfo"}).find("script").string
+    data_text = (
+        soup.find("div", attrs={"id": "fundinfo"}).find("script").string
+    )
     data_content = [
         item.split("|")
         for item in data_text.split("var")[6]
@@ -37,28 +39,28 @@ def fund_rating_all() -> pd.DataFrame:
         "简称",
         "类型",
         "基金经理",
-        "_",
+        "-",
         "基金公司",
-        "_",
+        "-",
         "5星评级家数",
-        "上海证券",
-        "上海证券-较上期",
+        "-",
+        "-",
         "招商证券",
-        "招商证券-较上期",
+        "-",
+        "上海证券",
+        "-",
+        "-",
+        "-",
         "济安金信",
-        "济安金信-较上期",
-        "_",
-        "_",
-        "_",
-        "_",
+        "-",
         "手续费",
-        "_",
-        "_",
-        "_",
-        "_",
-        "_",
-        "_",
-        "_",
+        "-",
+        "-",
+        "-",
+        "-",
+        "-",
+        "-",
+        "-",
     ]
     temp_df = temp_df[
         [
@@ -74,10 +76,15 @@ def fund_rating_all() -> pd.DataFrame:
             "类型",
         ]
     ]
+    temp_df["5星评级家数"] = pd.to_numeric(temp_df["5星评级家数"])
+    temp_df["上海证券"] = pd.to_numeric(temp_df["上海证券"])
+    temp_df["招商证券"] = pd.to_numeric(temp_df["招商证券"])
+    temp_df["济安金信"] = pd.to_numeric(temp_df["济安金信"])
+    temp_df["手续费"] = pd.to_numeric(temp_df["手续费"].str.strip("%")) / 100
     return temp_df
 
 
-def fund_rating_sh(date: str = '20200930') -> pd.DataFrame:
+def fund_rating_sh(date: str = "20200930") -> pd.DataFrame:
     """
     天天基金网-基金评级-上海证券评级
     http://fund.eastmoney.com/data/fundrating_3.html
@@ -87,14 +94,19 @@ def fund_rating_sh(date: str = '20200930') -> pd.DataFrame:
     url = "http://fund.eastmoney.com/data/fundrating_3.html"
     r = requests.get(url)
     soup = BeautifulSoup(r.text, "lxml")
-    date_list = [item['value'] for item in soup.find('select', attrs={'id': 'rqoptions'})]
-    date_format = '-'.join([date[:4], date[4:6], date[6:]])
+    date_list = [
+        item["value"]
+        for item in soup.find("select", attrs={"id": "rqoptions"})
+    ]
+    date_format = "-".join([date[:4], date[4:6], date[6:]])
     if date_format not in date_list:
-        return '请访问 http://fund.eastmoney.com/data/fundrating_3.html 获取查询日期'
+        return "请访问 http://fund.eastmoney.com/data/fundrating_3.html 获取查询日期"
     url = f"http://fund.eastmoney.com/data/fundrating_3_{'-'.join([date[:4], date[4:6], date[6:]])}.html"
     r = requests.get(url)
     soup = BeautifulSoup(r.text, "lxml")
-    data_text = soup.find("div", attrs={"id": "fundinfo"}).find("script").string
+    data_text = (
+        soup.find("div", attrs={"id": "fundinfo"}).find("script").string
+    )
     data_content = [
         item.split("|")
         for item in data_text.split("var")[1]
@@ -153,7 +165,7 @@ def fund_rating_sh(date: str = '20200930') -> pd.DataFrame:
     return temp_df
 
 
-def fund_rating_zs(date: str = '20201030') -> pd.DataFrame:
+def fund_rating_zs(date: str = "20201030") -> pd.DataFrame:
     """
     天天基金网-基金评级-招商证券评级
     http://fund.eastmoney.com/data/fundrating_2.html
@@ -163,14 +175,19 @@ def fund_rating_zs(date: str = '20201030') -> pd.DataFrame:
     url = "http://fund.eastmoney.com/data/fundrating_2.html"
     r = requests.get(url)
     soup = BeautifulSoup(r.text, "lxml")
-    date_list = [item['value'] for item in soup.find('select', attrs={'id': 'rqoptions'})]
-    date_format = '-'.join([date[:4], date[4:6], date[6:]])
+    date_list = [
+        item["value"]
+        for item in soup.find("select", attrs={"id": "rqoptions"})
+    ]
+    date_format = "-".join([date[:4], date[4:6], date[6:]])
     if date_format not in date_list:
-        return '请访问 http://fund.eastmoney.com/data/fundrating_2.html 获取查询日期'
+        return "请访问 http://fund.eastmoney.com/data/fundrating_2.html 获取查询日期"
     url = f"http://fund.eastmoney.com/data/fundrating_2_{'-'.join([date[:4], date[4:6], date[6:]])}.html"
     r = requests.get(url)
     soup = BeautifulSoup(r.text, "lxml")
-    data_text = soup.find("div", attrs={"id": "fundinfo"}).find("script").string
+    data_text = (
+        soup.find("div", attrs={"id": "fundinfo"}).find("script").string
+    )
     data_content = [
         item.split("|")
         for item in data_text.split("var")[1]
@@ -224,7 +241,7 @@ def fund_rating_zs(date: str = '20201030') -> pd.DataFrame:
     return temp_df
 
 
-def fund_rating_ja(date: str = '20200930') -> pd.DataFrame:
+def fund_rating_ja(date: str = "20200930") -> pd.DataFrame:
     """
     天天基金网-基金评级-济安金信评级
     http://fund.eastmoney.com/data/fundrating_4.html
@@ -234,14 +251,19 @@ def fund_rating_ja(date: str = '20200930') -> pd.DataFrame:
     url = "http://fund.eastmoney.com/data/fundrating_4.html"
     r = requests.get(url)
     soup = BeautifulSoup(r.text, "lxml")
-    date_list = [item['value'] for item in soup.find('select', attrs={'id': 'rqoptions'})]
-    date_format = '-'.join([date[:4], date[4:6], date[6:]])
+    date_list = [
+        item["value"]
+        for item in soup.find("select", attrs={"id": "rqoptions"})
+    ]
+    date_format = "-".join([date[:4], date[4:6], date[6:]])
     if date_format not in date_list:
-        return '请访问 http://fund.eastmoney.com/data/fundrating_4.html 获取查询日期'
+        return "请访问 http://fund.eastmoney.com/data/fundrating_4.html 获取查询日期"
     url = f"http://fund.eastmoney.com/data/fundrating_4_{'-'.join([date[:4], date[4:6], date[6:]])}.html"
     r = requests.get(url)
     soup = BeautifulSoup(r.text, "lxml")
-    data_text = soup.find("div", attrs={"id": "fundinfo"}).find("script").string
+    data_text = (
+        soup.find("div", attrs={"id": "fundinfo"}).find("script").string
+    )
     data_content = [
         item.split("|")
         for item in data_text.split("var")[1]
@@ -300,11 +322,11 @@ if __name__ == "__main__":
     fund_rating_all_df = fund_rating_all()
     print(fund_rating_all_df)
 
-    fund_rating_sh_df = fund_rating_sh(date='20200930')
+    fund_rating_sh_df = fund_rating_sh(date="20200930")
     print(fund_rating_sh_df)
 
-    fund_rating_zs_df = fund_rating_zs(date='20201030')
+    fund_rating_zs_df = fund_rating_zs(date="20201030")
     print(fund_rating_zs_df)
 
-    fund_rating_ja_df = fund_rating_ja(date='20200930')
+    fund_rating_ja_df = fund_rating_ja(date="20200930")
     print(fund_rating_ja_df)