+++ docs/changelog.md
@@ -29,8 +29,12 @@
 
 ## 更新说明
 
+1.6.79 fix: fix fund_rating_all interface
+
+    1. 修复 fund_rating_all 接口，字段与目标网站对齐
+
 1.6.78 fix: fix stock_telegraph_cls interface
-    
+
     1. 移除 pycryptodomex 依赖
     2. 移除冗余代码
 
@@ -1015,6 +1019,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.6.79 fix: fix fund_rating_all interface
 
 1.6.78 fix: fix stock_telegraph_cls interface